requirements.txt
Cython==0.29.34
cytoolz==0.11.0
dlib==19.24.0
easydict==1.10
einops==0.3.2
gradio==3.32.0
gradio_client==0.2.5
huggingface-hub==0.14.1
omegaconf==2.2.3
onnx==1.12.0
onnxruntime==1.9.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.4.60
opencv-python-headless==4.7.0.72
pandas==1.4.4
Pillow==9.2.0
pytorch-lightning==1.8.1
torch==1.12.1
torchvision==0.13.1